Move TLS module to separate crate
