+++ crates/common/axum_tls/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "axum_tls"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+axum = { workspace = true }
+axum-server = { workspace = true }
+camino = { workspace = true }
+futures = { workspace = true }
+pin-project = { workspace = true }
+rustls = { workspace = true }
+rustls-pemfile = { workspace = true }
+tokio = { workspace = true }
+tokio-rustls = { workspace = true }
+tower = { workspace = true }
+tracing = { workspace = true }
+x509-parser = { workspace = true }
