+++ crates/extensions/c8y_auth_proxy/Cargo.toml
@@ -13,12 +13,13 @@ repository = { workspace = true }
 anyhow = { workspace = true }
 axum = { workspace = true, features = ["macros"] }
 axum-server = { workspace = true }
+axum_tls = { workspace = true }
 c8y_http_proxy = { workspace = true }
 camino = { workspace = true }
 futures = { workspace = true }
 hyper = { workspace = true }
-rcgen = { workspace = true }
 pin-project = { workspace = true }
+rcgen = { workspace = true }
 reqwest = { workspace = true, features = ["stream"] }
 rustls = { workspace = true }
 rustls-pemfile = { workspace = true }
