+++ crates/extensions/c8y_auth_proxy/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod actor;
 mod body;
 mod server;
-
-mod tls;
 mod tokens;
 pub mod url;
